Refactor the code
Created new class SpecificationProcessor responsible for creating specification for tests.
